requirements.txt
altgraph==0.17.4
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
hupper==1.12.1
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
PasteDeploy==3.1.0
pefile==2023.2.7
plaster==1.1.2
plaster-pastedeploy==1.0.1
plyer==2.1.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
pyramid==2.0.2
pyramid_jinja2==2.10.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
translationstring==1.4
uritemplate==4.1.1
urllib3==2.2.2
venusian==3.1.0
WebOb==1.8.7
zope.deprecation==5.0
zope.interface==6.4.post2
